refac: tests refactored, now only 1 test run both test cases. #2391
