+++ tests/integration/test_logcollector/test_localfile/data/invalid_wazuh_conf.yaml
@@ -1,10 +1,12 @@
 ---
 - tags:
-  - test_invalid_localfile_config_no_log_format
+  - test_invalid_wazuh_conf
   apply_to_modules:
-  - test_invalid_localfile_config_no_log_format
+  - test_invalid_wazuh_conf
   sections:
     - section: localfile
       elements:
+      - log_format:
+          value: LOG_FORMAT
       - location:
           value: LOCATION
